netti-monitor/requirements.txt
cloudscraper==1.2.71
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.37.0
fake-useragent==2.2.0
undetected-chromedriver==3.5.5
setuptools==80.9.0
websockets==15.0.1
